refactor 지연 전파 facade로 꺼냄
