+++ src/main/java/com/ideality/coreflow/project/command/application/service/impl/TaskServiceImpl.java
@@ -194,167 +194,13 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    @Transactional
-    public DelayInfoDTO delayAndPropagate(Long taskId, Integer delayDays, boolean isSimulate) {
-        log.info("taskID: " + taskId);
-        Map<Long, Integer> visited = new HashMap<>();
-        Map<Long, Integer> realDelayed = new HashMap<>();
-        Queue<DelayNodeDTO> queue = new LinkedList<>();
-        Integer count = 0;
-        Set<LocalDate> holidays = holidayQueryService.getHolidays().stream()
-                .map(HolidayQueryDto::getDate).collect(Collectors.toSet());
-
-        // 초기 태스크 조회
-
-        // 지연된 태스크 ID를 추적할 Set
-        Set<Long> delayedTaskIds = new HashSet<>();
-
-        // 초기 태스크 처리
-        Work startTask = taskRepository.findById(taskId).orElseThrow(() -> new BaseException(TASK_NOT_FOUND));
-        Project project = projectRepository.findById(startTask.getProjectId()).orElseThrow(()->new BaseException(PROJECT_NOT_FOUND));
-
-        if (isSimulate) {
-            em.detach(startTask);
-            em.detach(project);
-        }
-
-        LocalDate projectEndExpect = project.getEndExpect();
-        LocalDate originProjectEndExpect = projectEndExpect;
-        System.out.println("projectEndExpect = " + projectEndExpect);
-        
-        int delayDaysByTask = 0;
-
-        // 지연 전파
-        queue.offer(new DelayNodeDTO(taskId, delayDays));
-        int startTaskDelay = Math.abs(delayDays - startTask.getSlackTime());
-
-        while (!queue.isEmpty()) {
-            DelayNodeDTO currentNode = queue.poll();
-
-            // 현재 taskId에 대해 visited보다 작은 지연일일 경우 스킵
-            Integer visitedDelay = visited.get(currentNode.getTaskId());
-            if (visitedDelay != null && visitedDelay > currentNode.getDelayDays()) {
-                continue;
-            }
-
-            // 현재 태스크 지연 설정
-            Work currentTask = taskRepository.findById(currentNode.getTaskId()).orElseThrow(() -> new BaseException(TASK_NOT_FOUND));
-            if (isSimulate) {
-                em.detach(currentTask);
-            }
-
-            // 현재 태스크의 지연일
-            int delayToApply = currentNode.getDelayDays();
-
-            // 현재 태스크의 슬랙타임 및 지연일 설정
-            if (currentTask.getSlackTime() >= delayToApply) {
-                currentTask.setSlackTime(currentTask.getSlackTime() - delayToApply);
-                if (Objects.equals(currentTask.getId(), taskId)) {
-                    currentTask.setDelayDays(delayToApply);
-                }
-                currentTask.setEndExpect(currentTask.getEndExpect().plusDays(
-                        calculateDelayExcludingHolidays(currentTask.getEndExpect(), delayDays, holidays)
-                ));
-            } else {
-                log.info("슬랙타임 내에서 해결 실패");
-                count++;
-                int realDelay = delayToApply - currentTask.getSlackTime();
-                realDelayed.put(currentTask.getId(), realDelay);
-                System.out.println("delayToApply = " + delayToApply);
-                System.out.println("slackTime = " + currentTask.getSlackTime());
-                System.out.println("realDelay: " + realDelay);
-                System.out.println("taskId = " + currentTask.getId());
-                if (Objects.equals(currentTask.getId(), taskId)) {
-                    currentTask.setDelayDays(currentTask.getDelayDays() + realDelay);   // 지연일 업데이트는 초기노드만 수정 필요
-                }
-                currentTask.setSlackTime(0);
-
-                // 예상 마감일이 변경될 경우만 추적
-                LocalDate originalEndExpect = currentTask.getEndExpect();
-
-                // 현재 태스크와 세부일정 예상 마감일 미루기
-                projectEndExpect=delayTask(currentTask, realDelay, holidays, projectEndExpect, isSimulate);
-
-                // 예상 종료일이 변경된 경우에만 태스크ID 추가
-                if (!currentTask.getEndExpect().equals(originalEndExpect)) {
-                    delayedTaskIds.add(currentTask.getId());  // 실제로 마감일이 변경된 태스크만 추가
-                    // 지연된 태스크에 참여한 인원에게 알림 보내기
-                    List<Long> participants = participantMapper.findParticipantsByTaskId(currentTask.getId());
-                    String contents = "태스크 [" + startTask.getName() + "]가 지연되어 ["+ currentTask.getName() +"]의 예상마감일이 변경되었습니다!";
-                    for (Long userId : participants) {
-                        notificationService.sendNotification(userId, contents, currentTask.getId(), TargetType.WORK);
-                    }
-                }
-
-                // 다음 노드에 realDelay를 전파
-                List<Long> nextTaskIds = relationQueryService.findNextTaskIds(currentNode.getTaskId());
-
-                for (Long nextTaskId  : nextTaskIds) {
-                    // 다음 노드에 저장된 지연일
-                    Integer storedDelay = visited.get(nextTaskId);
-                    if (storedDelay == null || realDelay > storedDelay) {
-                        visited.put(nextTaskId, realDelay);
-                        queue.offer(new DelayNodeDTO(nextTaskId, realDelay));
-                    }
-                }
-            }
-            if (!isSimulate) {
-                taskRepository.save(currentTask);
-            }
-        }
-
-        // 프로젝트 예상 마감일 업데이트
-        if (project.getEndExpect().isBefore(projectEndExpect)) {
-            Long projectDelay = ChronoUnit.DAYS.between(project.getEndExpect(), projectEndExpect)
-                    -holidayQueryService.countHolidaysBetween(project.getEndExpect(), projectEndExpect);
-            // 프로젝트 지연일수 업데이트
-            // 해당 지연으로 밀린 프로젝트 지연일
-            delayDaysByTask = Math.toIntExact(projectDelay);
-            project.setEndExpect(projectEndExpect);
-            project.setDelayDays(project.getDelayDays() + delayDaysByTask);
-            
-
-            if (!isSimulate) {
-                projectRepository.save(project);
-            }
-
-            // 프로젝트 마감일이 변경되었으면, 프로젝트 디렉터에게 알림 보내기
-            Long directorUserId = participantMapper.findDirectorByProjectId(project.getId());
-            String directorContent = "프로젝트 [" + project.getName() + "]의 예상 마감일이 지연되었습니다!";
-            notificationService.sendNotification(directorUserId, directorContent, project.getId(), TargetType.PROJECT);
-        }
-
-        visited.put(startTask.getId(), startTaskDelay);
-
-        // 지연된 태스크 ID 출력
-        System.out.println("지연된 태스크 ID들: " + delayedTaskIds);
-
-        // 반환할 값
-
-        log.info("기존 프로젝트 예상 마감일: {}", originProjectEndExpect);
-        log.info("전체 지연일: {}", delayDaysByTask);
-        log.info("지연 반영 된 프로젝트 마감일: {}", projectEndExpect);
-        log.info("결재 요청 온 현재 태스크 지연일:{}", startTaskDelay);
-        log.info("지연된 태스크 갯수:{}", count);
-        log.info("영향 받은 태스크 id별 지연일:{}", visited);
-
-
-        return DelayInfoDTO.builder()
-                .originProjectEndExpect(originProjectEndExpect)
-                .newProjectEndExpect(projectEndExpect)
-                .delayDaysByTask(delayDaysByTask)
-                .taskCountByDelay(count)
-                .delayDaysByTaskId(realDelayed)
-                .build();
-    }
-
-    @Override
     public String findTaskNameById(long taskId) {
         return taskRepository.findById(taskId).orElseThrow(() -> new BaseException(TASK_NOT_FOUND)).getName();
     }
 
 
-    private LocalDate delayTask(Work task, Integer delayDays, Set<LocalDate> holidays, LocalDate projectEndExpect, boolean isSimulate) {
+    @Override
+    public LocalDate delayTask(Work task, Integer delayDays, Set<LocalDate> holidays, LocalDate projectEndExpect, boolean isSimulate) {
         if (task.getStatus() == Status.PENDING) {
             task.setStartExpect(task.getStartExpect().plusDays(
                     calculateDelayExcludingHolidays(task.getStartExpect(), delayDays, holidays)
@@ -389,7 +235,8 @@ public class TaskServiceImpl implements TaskService {
         return projectEndExpect;
     }
 
-    private void delayWork(Work work, Integer delayDays, Set<LocalDate> holidays, boolean isSimulate) {
+    @Override
+    public void delayWork(Work work, Integer delayDays, Set<LocalDate> holidays, boolean isSimulate) {
         if (work.getStatus() == Status.PENDING) {
             work.setStartExpect(work.getStartExpect().plusDays(
                     calculateDelayExcludingHolidays(work.getStartExpect(), delayDays, holidays)
@@ -405,7 +252,8 @@ public class TaskServiceImpl implements TaskService {
         }
     }
 
-    private int calculateDelayExcludingHolidays(LocalDate startDate, Integer delayDays, Set<LocalDate> holidays) {
+    @Override
+    public int calculateDelayExcludingHolidays(LocalDate startDate, Integer delayDays, Set<LocalDate> holidays) {
         int addedDays = 0;
         int workingDays = 0;
         LocalDate date = startDate;
@@ -421,6 +269,11 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
+    public void taskSave(Work currentTask) {
+        taskRepository.save(currentTask);
+    }
+
+    @Override
     @Transactional
     public Long modifyTaskDetail(RequestModifyTaskDTO requestModifyTaskDTO, Long taskId) {
         Work modifyTask =
@@ -430,5 +283,10 @@ public class TaskServiceImpl implements TaskService {
                 requestModifyTaskDTO.getStartExpect(),
                 requestModifyTaskDTO.getEndExpect());
         return modifyTask.getId();
+    }
+
+    @Override
+    public Work findById(Long taskId) {
+        return taskRepository.findById(taskId).orElseThrow(() -> new BaseException(TASK_NOT_FOUND));
     }
 }
