+++ src/main/resources/application.yml
@@ -1,3 +1,3 @@
 spring:
   profiles:
-    active: deploy
+    active: dev
